+++ OUDS/Core/Components/Sources/Controls/TextInput/OUDSTextInput.swift
@@ -172,9 +172,9 @@ import SwiftUI
 ///
 /// ![A text input component in light and dark modes with Orange theme](component_textInput_Orange)
 ///
-/// ### Orange Business Tools
+/// ### Orange Compact
 ///
-/// ![A text input component in light and dark modes with Orange Business Tools theme](component_textInput_OrangeBusinessTools)
+/// ![A text input component in light and dark modes with Orange Compact theme](component_textInput_OrangeCompact)
 ///
 /// ### Sosh
 ///