+++ OUDS/Core/Themes/OrangeCompact/Sources/Providers/ComponentTokens/OrangeCompactThemeLinkComponentTokensProvider.swift
@@ -1,0 +1,74 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSFoundations
+import OUDSThemesContract
+
+// swiftlint:disable type_name
+
+/// A class which wraps all **component  tokens of links** for *link* objects like `OUDSLink`.
+/// Contains also references to semantic tokens providers so as to be able to use them to define the component tokens.
+/// This provider should be integrated as a `AllLinkComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
+/// all tokens to the users. It helps users to override some of the tokens and assign them to an `OUDSTheme` implementation to use.
+/// It implements also the protocol `LinkComponentTokens` so as to expose the component tokens for *links* through any `OUDSTheme`.
+/// *Link* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`),
+/// colors (from `AllColorSemanticTokensProvider`) and spaces (from `AllSpaceSemanticTokensProvider`).
+///
+/// - Since: 0.17.0
+final class OrangeCompactThemeLinkComponentTokensProvider: AllLinkComponentTokensProvider {
+
+    /// Provider of size semantic tokens to use for link sizes
+    let sizes: AllSizeSemanticTokensProvider
+
+    /// Provider of color semantic tokens to use for link colors
+    let colors: AllColorSemanticTokensProvider
+
+    /// Provider of spaces semantic tokens to use for link spaces
+    let spaces: AllSpaceSemanticTokensProvider
+
+    #if DEBUG
+    nonisolated(unsafe) private static var instanceCount: Int = 0
+    #endif
+
+    /// Defines a provider of component tokens dedicated to `OUDSLink`
+    /// - Parameters:
+    ///    - sizes: Provider for size semantic tokens. If nil, a default one will be used (``OrangeCompactThemeSizeSemanticTokensProvider``)
+    ///    - colors: Provider for color semantic tokens. If nil, a default one will be used (``OrangeCompactThemeColorSemanticTokensProvider``)
+    ///    - spaces: Provider for space semantic tokens. If nil, a default one will be used (``OrangeCompactThemeSpaceSemanticTokensProvider``)
+    init(sizes: AllSizeSemanticTokensProvider? = nil,
+         colors: AllColorSemanticTokensProvider? = nil,
+         spaces: AllSpaceSemanticTokensProvider? = nil)
+    {
+        OL.debug("Init of OrangeCompactThemeLinkComponentTokensProvider")
+        self.sizes = (sizes ?? OrangeCompactThemeSizeSemanticTokensProvider())
+        self.colors = (colors ?? OrangeCompactThemeColorSemanticTokensProvider())
+        self.spaces = (spaces ?? OrangeCompactThemeSpaceSemanticTokensProvider())
+        #if DEBUG
+        Self.instanceCount++
+        checkInstances(count: Self.instanceCount, for: "OrangeCompactThemeLinkComponentTokensProvider")
+        #endif
+    }
+
+    deinit {
+        #if DEBUG
+        Self.instanceCount--
+        #endif
+    }
+
+    // ଘ( ･ω･)_/ﾟ･:*:･｡☆
+    // Note: So as to help the integration of generated code produced by the tokenator
+    // the implementation of LinkComponentTokens is not here but in Core/Themes/OrangeCompact/Values/ComponentTokens/OrangeCompactTheme+LinkComponentTokens.swift
+    // This declaration of OrangeCompactThemeLinkComponentTokensProvider is here also to allow to write documentation.
+}
+
+// swiftlint:enable type_name