+++ OUDS/exported/OUDSSwiftUIOrange/Sources/Exported.swift
@@ -31,7 +31,7 @@
 @_exported import OUDSModules
 @_exported import OUDSThemesContract
 @_exported import OUDSThemesOrange
-@_exported import OUDSThemesOrangeBusinessTools
+@_exported import OUDSThemesOrangeCompact
 @_exported import OUDSTokensComponent
 @_exported import OUDSTokensRaw
 @_exported import OUDSTokensSemantic